+++ myplugin/content/admin/utils.py
@@ -1,0 +1,99 @@
+from openstack_dashboard.api import keystone
+import requests
+import logging
+from myplugin.content.api_endpoints import API_ENDPOINTS
+
+def get_username_from_id(request, user_id):
+    try:
+        user = keystone.user_get(request, user_id)
+        return user.name
+    except Exception:
+        return user_id
+
+
+
+def get_token_id(request):
+    """
+    Retrieves the token ID from the request object.
+    """
+    return getattr(getattr(request, "user", None), "token", None) and request.user.token.id
+
+def get_users(request):
+    """
+    Fetches app templates from the external API using a provided token ID.
+    """
+    token_id = get_token_id(request)
+    headers = {"X-Auth-Token": token_id}
+
+    try:
+        response = requests.get(API_ENDPOINTS['user_list'],
+                                headers=headers, timeout=10)
+        response.raise_for_status()
+        return response.json()
+    except requests.RequestException as e:
+        logging.error("Failed to fetch users: %s", e)
+        return []
+
+def get_roles(request):
+    """
+    Fetches app templates from the external API using a provided token ID.
+    """
+    token_id = get_token_id(request)
+    headers = {"X-Auth-Token": token_id}
+
+    try:
+        response = requests.get(API_ENDPOINTS['roles_list'],
+                                headers=headers, timeout=10)
+        response.raise_for_status()
+        return response.json()
+    except requests.RequestException as e:
+        logging.error("Failed to fetch roles: %s", e)
+        return []
+
+def get_user_details(request, user_id):
+    """
+    Fetches detailed user information for a given user_id using the external API.
+    """
+    token_id = get_token_id(request)
+    headers = {"X-Auth-Token": token_id}
+    url = f"{API_ENDPOINTS['user_list']}{user_id}"
+
+    try:
+        response = requests.get(url, headers=headers, timeout=10)
+        response.raise_for_status()
+        return response.json()
+    except requests.RequestException as e:
+        logging.error("Failed to fetch user details for user_id %s: %s", user_id, e)
+        return {}
+
+def get_app_templates_to_approve(request):
+    """
+    Fetches app templates to approve from the external API using a provided token ID.
+    """
+    token_id = get_token_id(request)
+    headers = {"X-Auth-Token": token_id}
+
+    try:
+        response = requests.get(API_ENDPOINTS['get_to_approve'],
+                                headers=headers, timeout=10)
+        response.raise_for_status()
+        return response.json()
+    except requests.RequestException as e:
+        logging.error("Failed to fetch app templates to approve: %s", e)
+        return []
+
+def get_app_templates(request):
+    """
+    Fetches app templates from the external API using a provided token ID.
+    """
+    token_id = get_token_id(request)
+    headers = {"X-Auth-Token": token_id}
+
+    try:
+        response = requests.get(API_ENDPOINTS['app_templates'],
+                                headers=headers, timeout=10)
+        response.raise_for_status()
+        return response.json()
+    except requests.RequestException as e:
+        logging.error("Failed to fetch app templates: %s", e)
+        return []
